Refs #212, Disabling tokens when logout, use redis to store disabled tokens.
