+++ teraserver/python/modules/FlaskModule/API/participant/ParticipantLogout.py
@@ -34,7 +34,7 @@ class ParticipantLogout(Resource):
             if 'Authorization' in request.headers:
                 scheme, old_token = request.headers['Authorization'].split(None, 1)
                 if scheme == 'OpenTera':
-                    LoginModule.participant_push_disabled_token(old_token)
+                    LoginModule.participant_add_disabled_token(old_token)
 
             return gettext("Participant logged out."), 200
         else:
