+++ teraserver/python/modules/FlaskModule/API/user/UserRefreshToken.py
@@ -69,7 +69,7 @@ class UserRefreshToken(Resource):
         # Put old token in disabled tokens
         scheme, old_token = request.headers['Authorization'].split(None, 1)
         if len(old_token) > 0:
-            LoginModule.user_push_disabled_token(old_token)
+            LoginModule.user_add_disabled_token(old_token)
 
         # Regenerate token, 30 minutes expiration
         user_token = current_user.get_token(token_key, expiration=60 * 30)
